refactor #2377: renamed injection to di & modules to objects
